+++ AzLocal.UpdateManagement/Tests/AzLocal.UpdateManagement.Tests.ps1
@@ -34,8 +34,8 @@ Describe 'Module: AzLocal.UpdateManagement' {
             $script:ModuleInfo | Should -Not -BeNullOrEmpty
         }
 
-        It 'Should have version 0.9.1' {
-            $script:ModuleInfo.Version | Should -Be '0.9.1'
+        It 'Should have version 0.9.10' {
+            $script:ModuleInfo.Version | Should -Be '0.9.10'
         }
 
         It 'Module version constants are in sync between .psm1 and .psd1' {